correct directories structure
