+++ battleship-server/src/test/java/com/epam/solid/nie/config/language/CommunicateProviderImplTest.java
@@ -16,11 +16,33 @@ public class CommunicateProviderImplTest {
     }
 
     @Test(dataProvider = "languages")
-    public void populateShould_FillCommunicatesMap(Language language) {
+    public void populateShould_FillMapWithCommunicates(Language language) throws Exception {
         communicateProvider.populate(language);
         assertFalse(communicateProvider.isMapEmpty());
     }
 
+    @DataProvider
+    public static Object[][] communicatesInEnglish() {
+        return new Object[][]{{Language.ENGLISH,"Welcome in my game.", Communicate.WELCOME},
+        };
+    }
+
+    @Test (dataProvider = "communicatesInEnglish")
+    public void shouldLoadCorrectCommunicates_English(Language language,String expectedCommunicate,Communicate communicate){
+        communicateProvider.populate(language);
+        assertEquals(expectedCommunicate,communicateProvider.getCommunicate(communicate));
+    }
+    @DataProvider
+    public static Object[][] communicatesInPolish() {
+        return new Object[][]{{Language.POLISH,"Witaj w  mojej grze.", Communicate.WELCOME},
+        };
+    }
+
+    @Test (dataProvider = "communicatesInPolish")
+    public void shouldLoadCorrectCommunicates_Polish(Language language,String expectedCommunicate,Communicate communicate){
+        communicateProvider.populate(language);
+        assertEquals(expectedCommunicate,communicateProvider.getCommunicate(communicate));
+    }
 
 
 }
