+++ battleship-server/src/test/java/com/epam/solid/nie/config/language/LanguageConfigurationImplTest.java
@@ -6,15 +6,14 @@ import static org.testng.Assert.*;
 
 public class LanguageConfigurationImplTest {
     private LanguageConfiguration languageConfiguration = new LanguageConfigurationImpl();
-
     @Test
-    public void shouldRead_Default_LanguageValue_FromFile() {
+    public void shouldRead_Default_LanguageValue_FromFile() throws Exception {
         Language currentLanguage = languageConfiguration.provide();
         assertTrue(currentLanguage.equals(Language.ENGLISH));
     }
 
     @Test
-    public void shouldLoad_AnyLanguage_FromFile() {
+    public void shouldLoad_AnyLanguage_FromFile() throws Exception {
         assertNotNull(languageConfiguration.provide());
     }
 }
